+++ mod-data-import-converter-storage-server/src/main/java/org/folio/dao/JobProfileDaoImpl.java
@@ -10,7 +10,7 @@ import org.springframework.stereotype.Component;
  * Data access object for {@link JobProfile}
  */
 @Component
-public class JobProfileDaoImpl extends AbstractProfileDao <JobProfile, JobProfileCollection> {
+public class JobProfileDaoImpl extends AbstractProfileDao<JobProfile, JobProfileCollection> {
 
   private static final String JOB_PROFILES_TABLE_NAME = "job_profiles";